pci_devices: run "info qtree" after the guest boots up
QEMU "info qtree" output depends on information generated by guest
firmware. One example is the bus number in PCI device address. QEMU
retrieves it from the PCI configuration space of the bridge which
creates the bus that a PCI device is attached to. So if test runs
"info qtree" before firmware finishes initializing pci devices, its
output can be incorrect.

The current code works around this timing issue by sleeping 5 seconds.
That is not reliable and doesn't work on my aarch64 system. So I'd
suggest to change the test to not start until it verifies the guest
has booted up. This will make the test execution time slightly longer
but is reliable.

Signed-off-by: Huan Xiong <huan.xiong@hxt-semitech.com>