live_snapshot_chain: Flush data to disk before running system_reset
QMP command 'system_reset' is issued after creating files in the guest. The
command has an effect similar to the physical reset button on a PC. But
filesystems may be left in an unclean state which causes the loss of data. Run
the command "sync" to flush the data to disk.

Signed-off-by: Ping Li <pingl@redhat.com>