qemu..smbios: Avoid defining extra pci_controllers
The extra pci_controllers are needed on q35 for hotplug purposes, but
this test even on q35 machine-type attempts to boot all supported
machine-types including i440fx. The i440fx requires additional PCIE
root-complex and this is not handled by Avocado-vt automatically and
would require complex handling only to get something that is not needed
by the test. Let's simply not create the additional controllers as we
don't need them.

Without this patch we are getting:

    "Can't add chassis slot, error -16"

because Avocado-vt does not define chassis for PCI busses (it does that
only for PCIE busses). Yes, it'd be possible to add support for extra
controllers for i440fx, but at this point it is not necessary and this
solution is way simpler.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>