qemu.migration: Add support for pre/post-migrate for reboot mig
Currently the "pre_migrate" and "post_migrate" is only supported in the
stock "migrate.py" test. This patch puts the related functions on top of
the module, adds the needed variables as arguments and uses those
methods from the migrate.with_reboot test.

Note it'd be much nicer to implement this in framework, but that would
take a lot of backward-compatible code, therefor this looks like a
bearable compromise.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>