qemu..migration: Run pre_migrate before each iteration
Previously the "pre_migrate" function was only executed before the first
iteration, but at least for mig_set_speed it's required before each
iteration.

As for the "check_dma" function, it also makes sense to run it for each
iteration and code-wise should work well, but because of outdated
code-path I have only tested it with slighly modified version (without
dmesg check on virtio_blk).

Note the migration_with_speed_measurement test is not affected as it's
located in a different file.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>