iozone_windows: Reset disk letter if disk letter change after iozone
For iozone_windows case will be called as a sub-case by other cases.
after system_reset during iozone test on the vioscsi/viostor data disk,
found the the disk letter of iozone tested disk changed.Before iozone
test, disk letter is "I", after system_reset disk letter changed to "E".
It is a product bug but closed as WONTFIX, so need to workaround it in
automation.If found the disk letter changed, drop the origin disk letter
then reset the disk letter to the target letter.

Signed-off-by: Menghuan Li menli@redhat.com