plug_cdrom: special device /dev/sr0 does not exist
This commit fixes the issue that this testcase was having with linux
guests.
The root cause was that the testcase was trying to mount plugged devices
without previously checking if they were detected in /dev/...
Seemly, a similar issue had been reported for windows guests and another
test workflow was introduced in
commit #ec07974a7c9959bd9f667f775a11f3d441349101 for windows guests.
In this commit the same workflow is also adopted for linux guests. Now,
no matter what kind of OS the guest is running, the test will wait until
the device is available before is mounted.

Signed-off-by: Beñat Gartzia <bgartzia@redhat.com>