block_hotplug_in_pause: stg0_encrypted not deleted
Script was expecting deleted events coming back from QMP even for qemu
objects. Regarding that, only the block_del commands should be getting
an event back when finally deleted. Objects, instead, are directly
deleted, and so, no event is generated to know when they have been
deleted from the tree.
This commit applies changes so the script filters out the QObjects when
verifying on the QMP monitor out, and only waits for the DEVICE_DELETED
event when deleting block devices.

Signed-off-by: Beñat Gartzia <bgartzia@redhat.com>