multi_disk_random_hotplug: cant insert ehci device
The multi_disk_random_hotplug test was failing when trying to add a 6th
device into the ehci.0 bus. The reason behind it was that although
having another ehci bus named usb1.0 the by-default usb-tablet1 was
being attached into the bus named ehci.0 bus, used by the test for the
hotplug/unplug workflow.
I think that the issue should be fixed in the base.cfg. However that
could be risky since it wouldn't be affecting only to tests from this
provider. I'll open another issue in the avocado-vt repo to discuss
about the usb_bus config.

Signed-off-by: Beñat Gartzia <bgartzia@redhat.com>