usb: usb_ variables changed to usbdev_
usb parameters as usb_bus or usb_port were not being parsed as they
should be from virttest/qemu_devices/qcontainer.py. Due to that reason,
the key that the framework will look for in the parameter dict has been
changed to usbdev_bus or usbdev_port from now on. Also usb_type has been
changed to usbdev_type when referring to usb devices and not controllers
for consistency. That change caused that some test cases needed to be
updated. This commit adds such changes.

Signed-off-by: Beñat Gartzia <bgartzia@redhat.com>