vt_ulimit_nofile: workarounds deletion
I did some research over the test cases that were using the
vt_ulimit_nofile configuration parameter. The goal was to delete it from
tests using it to avoid the "Too many open files" error as a workaround,
since a fix for the root cause was merged a few weeks ago.
However, I only catched a test case using the parameter as a workaround.
The second deletion (in multi_disk_random_hotplug) is due to a redundant
assignment.
It looks like the rest of test cases using the vt_ulimit_nofile are
doing it with other goals in mind rather than avoiding getting an error
for the fd-leak when opening/closing sessions.

Signed-off-by: Beñat Gartzia <bgartzia@redhat.com>