bridge_qinq: Reduce log size
This is only a basic test to check extertypes, however tcpdump will
capture all packages and this makes the log size too huge. So reduce the
ping count to reduce the log size, also use "setsid" command to run
tcpdump in the background, to avoid dirty output when issuing "pkill
tcpdump".

Signed-off-by: Yihuang Yu <yihyu@redhat.com>