Cut out '-pci' from the namespace of virtio-rng device params
'virtio-rng-pci' is one type of virtio rng devices, the real name is not
unique, we have "virtio-rng-pci", "virtio-rng-pci-non-transitional",
"virtio-rng-pci-transitional", etc. And the device name in some
platforms is not the same. So we should not use the params namespace
"xxx_virtio-rng-pci", otherwise we need to configure the same virtio-rng
device with different parameters in diff platforms.

Signed-off-by: Yihuang Yu <yihyu@redhat.com>