block_diskcard: add sleep time between mklabel and mkpart
Sometimes it will fail to mkpart just after mklabel, it says that:
"Partition(s) 1 on /dev/sdb have been written, but we have been unable to inform the kernel of the change,
probably because it/they are in use. As a result, the old partition(s) will remain in use.
You should reboot now before making further changes."

Wait a little time can fix this problem.

Signed-off-by: Hu Shuai <hus.fnst@fujitsu.com>