memhp_threads: Update check point for prealloc-threads default
Some old version rhel8 hosts will create a thread when adding one
memory-backend-file default, but the thread creation is not due to the
setting of the qemu source code.
We can see that before qemu7.1 on rhel9, adding a memory-backend-file
default will not create one thread.
This was fixed after qemu7.1, anyway, Our test point is that after 7.1,
the default value of the thread is equal to the number of smps, so as to
achieve performance improvement.
So remove pre-7.1 default checks.

Signed-off-by: zhenyzha <zhenyzha@redhat.com>