tpm_bind_luks: Update FS options written to crypttab and fstab
It seems that a problem in the previous version, it causes the luks FS
sometimes not to mount after reboot, so correct the mount options in
crypttab and fstab. We should use UUID instead of disk path in crypttab
to avoid the disk letter changes, also relabel the mount point is a must
to avoid another failure.

Signed-off-by: Yihuang Yu <yihyu@redhat.com>