timer_rtc_sync: Ensure host/guest uses RTC in UTC
If the system is configured to read the RTC time in the local time zone,
the RTC time will not sync, timedatectl will also report a warning.

```
The system is configured to read the RTC time in the local time zone.
This mode cannot be fully supported. It will create various problems
with time zone changes and daylight saving time adjustments. The RTC
time is never updated, it relies on external facilities to maintain it.
If at all possible, use RTC in UTC by calling
'timedatectl set-local-rtc 0'.
```

Signed-off-by: Yihuang Yu <yihyu@redhat.com>