blockdev_snapshot_chains:update the method to get base format
In original design, miss this scenario: we define an image
format for base image that is different from the --imageformat
we defined in kar cmdline. In this scenario, when we do some
rebase for snapshot images, we will use an incorrect imageformat
for the base image, which will lead to "can't find baking image"
error when start guest with snapshot images.

Signed-off-by: Aihua Liang <aliang@redhat.com>