migration_virtio_mem_ignore_shared: creates new test case
Creates new case to test virtio-mem device migration with
x-ignore-shared capability enabled. It starts both guests,
source and destination ones, then puts some stress in the
first one during the migration process. Finally, does the
migration and checks that, only a little amount of memory
has been migrated because of the capability and the virtio-mem
device works as expected.

Signed-off-by: mcasquer <mcasquer@redhat.com>