thp_fragment: updates tool to v2 and extracts it to a new location
Updates the thp_fragment to the new version where the tool
works better in environments with a great amount of memory.
Also extracts this tool to the 'root deps' directory as it is
going to be used by a new provider script and the tarball will
no longer needed, this way the file can be reviewed easily by
the maintainers where new changes arrive and the extraction
step can be saved.

Signed-off-by: mcasquer <mcasquer@redhat.com>