nvdimm: removes epel dependecies and use DOC=n flag with make
As the documents compilation can be skipped using 'DOC=n' flag,
the pandoc package is no longer needed and in the same way is
the EPEL repository. Changes the approach to first call make
with the DOC=n flag, then build the tests with make test and
run the RUNTESTS.sh script to execute the tests.

Signed-off-by: mcasquer <mcasquer@redhat.com>