qemu_guest_agent: Add logic to get specific rpm from url
The Qemu rebase test requires the latest qemu-guest-agent package,
but the guest would not install the latest version of the series
of Qemu at the beginning. sometimes will be delayed for a while.
Plus I have to run qga test frequently in the future, so it's
necessary to add a piece of  logic to install the specific
package of qga from url.

Signed-off-by: Dehan Meng <demeng@redhat.com>