Update the method to check driver was installed
For currently win11 and win2025 always fail with
windows update case, the root cause is that:
For INF isolation all win11 drivers are moved to DriverStore
(C:\WINDOWS\system32\DriverStore\FileRepository).
Each installation creates a new directory in
the DriverStore folder depending on the driver hash.
While for win10 drivers, there are no change.
So change the method to check whether the driver was
installed. After the test, all guests work normally.

Signed-off-by: menli <menli@redhat.com>