blockdev_commit_auto_readonly:not set size for snapshot image
Snapshot image size set to 20G, but under some situation,
we start VM with system image size not equal to 20G, so
when we do the snapshot/commit with a snapshot image that
smaller that system image size, the case will fail with
"no ipv4 address" issue. So correct it to let the snapshot
image size follow the default system image size.

Signed-off-by: Aihua Liang <aliang@redhat.com>